Extract the shared spawn-loop-sleep pattern into monitor::run_periodic
auto_heal_monitor.rs, auto_stop_monitor.rs, git_status_monitor.rs,
disk_space_monitor.rs, and tls_expiry_monitor.rs each reimplemented
the same std::thread::spawn(move || loop { check(); sleep(interval) })
loop, with now_secs() copied byte-for-byte into three of them. All
five now call the same monitor::run_periodic (which also supports
auto-heal's startup grace delay) and monitor::now_secs.
