+++ src-tauri/src/git_status_monitor.rs
@@ -11,11 +11,7 @@ static LAST_ALERT: LazyLock<Mutex<HashMap<String, i64>>> =
     LazyLock::new(|| Mutex::new(HashMap::new()));
 
 pub fn start(app: &tauri::AppHandle) {
-    let app = app.clone();
-    std::thread::spawn(move || loop {
-        check(&app);
-        std::thread::sleep(CHECK_INTERVAL);
-    });
+    crate::monitor::run_periodic(app, CHECK_INTERVAL, Duration::ZERO, check);
 }
 
 fn check(app: &tauri::AppHandle) {
@@ -57,7 +53,7 @@ fn check(app: &tauri::AppHandle) {
 }
 
 fn should_notify(site_id: &str) -> bool {
-    let now = now_secs();
+    let now = crate::monitor::now_secs();
     let mut last_alert = LAST_ALERT.lock().unwrap_or_else(|error| error.into_inner());
     match last_alert.get(site_id) {
         Some(&last) if now - last < RENOTIFY_INTERVAL_SECS => false,
@@ -87,11 +83,4 @@ fn notify(
         )
     };
     crate::notifications::send(app, "git-status", "LS Panel", &body);
-}
-
-fn now_secs() -> i64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs() as i64
 }
