+++ src-tauri/src/tls_expiry_monitor.rs
@@ -11,11 +11,7 @@ const RENOTIFY_INTERVAL_MS: i64 = 24 * 60 * 60 * 1000;
 static LAST_ALERT_AT: AtomicI64 = AtomicI64::new(0);
 
 pub fn start(app: &tauri::AppHandle) {
-    let app = app.clone();
-    std::thread::spawn(move || loop {
-        check(&app);
-        std::thread::sleep(CHECK_INTERVAL);
-    });
+    crate::monitor::run_periodic(app, CHECK_INTERVAL, Duration::ZERO, check);
 }
 
 fn check(app: &tauri::AppHandle) {
